add prom metrics
