+++ infra/k8s/stock-ticker-watcher/values.yaml
@@ -82,3 +82,5 @@ tolerations: []
 affinity: {}
 
 apiKeySecret: "api-key-secret"
+
+exposePrometheusMetrics: false # Enable ServiceMonitor for Prometheus
